# Optimize reading numbers by removing redundant search for end of number