--- a/src/System.Text.JsonLab/System/Text/Json/JsonReader.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonReader.cs
@@ -735,24 +735,20 @@ namespace System.Text.JsonLab
             else if ((uint)(marker - '0') <= '9' - '0')
             {
                 ReadOnlySequence<byte> sequence = reader.Sequence.Slice(reader.Position);
-                Value = sequence.IsSingleSegment ? sequence.First.Span : sequence.ToArray();
+                Value = GetNumber(sequence.IsSingleSegment ? sequence.First.Span : sequence.ToArray());
                 ValueType = JsonValueType.Number;
-                ValidateNumber(Value);
-                int length = reader.UnreadSpan.Length;
-                reader.Advance(length);
-                Index += length;
+                reader.Advance(Value.Length);
+                Index += Value.Length;
             }
             else if (marker == '-')
             {
                 //TODO: Is this a valid check?
                 if (reader.End) JsonThrowHelper.ThrowJsonReaderException(ref this);
                 ReadOnlySequence<byte> sequence = reader.Sequence.Slice(reader.Position);
-                Value = sequence.IsSingleSegment ? sequence.First.Span : sequence.ToArray(); //TODO: any way to avoid the allocation?
+                Value = GetNumber(sequence.IsSingleSegment ? sequence.First.Span : sequence.ToArray()); //TODO: any way to avoid the allocation?
                 ValueType = JsonValueType.Number;
-                ValidateNumber(Value);
-                int length = reader.UnreadSpan.Length;
-                reader.Advance(length);
-                Index += length;
+                reader.Advance(Value.Length);
+                Index += Value.Length;
             }
             else if (marker == 'f')
             {
@@ -793,12 +789,11 @@ namespace System.Text.JsonLab
             }
             else if ((uint)(marker - '0') <= '9' - '0' || marker == '-')
             {
-                Value = _buffer;
+                Value = GetNumber(_buffer);
                 ValueType = JsonValueType.Number;
-                ValidateNumber(Value);
-                Index += _buffer.Length;
+                Index += Value.Length;
 #if UseInstrumented
-                _position += _buffer.Length;
+                _position += Value.Length;
 #endif
             }
             else if (marker == 'f')
@@ -839,27 +834,16 @@ namespace System.Text.JsonLab
 
             Value = span;
             ValueType = JsonValueType.Number;
-            ValidateNumber(Value);
+            GetNumber(Value);
         }
 #endif
         private void ConsumeNumberUtf8()
         {
-            int idx = _buffer.Slice(Index).IndexOfAny(JsonConstants.Delimiters);
-            if (idx == -1)
-            {
-#if UseInstrumented
-                throw new JsonReaderException($"Expected a delimiter that indicates end of number, but didn't find one.", _lineNumber, _position);
-#else
-                JsonThrowHelper.ThrowJsonReaderException(ref this);
-#endif
-            }
-
-            Value = _buffer.Slice(Index, idx);
             ValueType = JsonValueType.Number;
-            ValidateNumber(Value);
-            Index += idx;
+            Value = GetNumber(_buffer.Slice(Index));
+            Index += Value.Length;
 #if UseInstrumented
-            _position += idx;
+            _position += Value.Length;
 #endif
         }
 
@@ -1368,96 +1352,6 @@ namespace System.Text.JsonLab
         }
 
 #if UseInstrumented
-        // https://tools.ietf.org/html/rfc7159#section-6
-        //TODO: Investigate optimizations
-        private void ValidateNumber(ReadOnlySpan<byte> data)
-        {
-            Debug.Assert(data.Length > 0);
-
-            int i = 0;
-            byte nextByte = data[i++];
-
-            if (nextByte == '-')
-            {
-                if (i >= data.Length)
-                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
-                nextByte = data[i++];
-            }
-
-            Debug.Assert(nextByte >= '0' && nextByte <= '9');
-
-            if (nextByte == '0')
-            {
-                if (i < data.Length)
-                {
-                    nextByte = data[i];
-                    if (nextByte != '.' && nextByte != 'E' && nextByte != 'e')
-                        throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected '.' or 'E' or 'e'.", _lineNumber, _position);
-                    i++;
-                }
-                else
-                    return;
-            }
-            else
-            {
-                for (; i < data.Length; i++)
-                {
-                    nextByte = data[i];
-                    if ((uint)(nextByte - '0') > '9' - '0')
-                        break;
-                }
-                if (i >= data.Length)
-                    return;
-                if (nextByte != '.' && nextByte != 'E' && nextByte != 'e')
-                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected '.' or 'E' or 'e'.", _lineNumber, _position);
-                i++;
-            }
-
-            Debug.Assert(nextByte == '.' || nextByte == 'E' || nextByte == 'e');
-
-            if (nextByte == '.')
-            {
-                if (i >= data.Length)
-                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
-
-                for (; i < data.Length; i++)
-                {
-                    nextByte = data[i];
-                    if ((uint)(nextByte - '0') > '9' - '0')
-                        break;
-                }
-                if (i >= data.Length)
-                    return;
-                if (nextByte != 'E' && nextByte != 'e')
-                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected 'E' or 'e'.", _lineNumber, _position);
-                i++;
-            }
-
-            Debug.Assert(nextByte == 'E' || nextByte == 'e');
-
-            if (i >= data.Length)
-                throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
-
-            nextByte = data[i];
-            if (nextByte == '+' || nextByte == '-')
-            {
-                i++;
-                if (i >= data.Length)
-                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
-
-            }
-
-            for (; i < data.Length; i++)
-            {
-                nextByte = data[i];
-                if ((uint)(nextByte - '0') > '9' - '0')
-                    break;
-            }
-
-            if (i < data.Length)
-                throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
-        }
-
         private void AdjustLineNumber(ReadOnlySpan<byte> span)
         {
             //TODO: Avoid redoing first IndexOf search
@@ -1471,12 +1365,14 @@ namespace System.Text.JsonLab
             }
             _position = span.Length + 1;
         }
-#else
+#endif
+
         // https://tools.ietf.org/html/rfc7159#section-6
-        //TODO: Investigate optimizations
-        private void ValidateNumber(ReadOnlySpan<byte> data)
+        private ReadOnlySpan<byte> GetNumber(ReadOnlySpan<byte> data)
         {
             Debug.Assert(data.Length > 0);
+
+            ReadOnlySpan<byte> delimiters = JsonConstants.Delimiters;
 
             int i = 0;
             byte nextByte = data[i++];
@@ -1484,8 +1380,19 @@ namespace System.Text.JsonLab
             if (nextByte == '-')
             {
                 if (i >= data.Length)
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
+
                 nextByte = data[i++];
+                if (delimiters.IndexOf(nextByte) != -1)
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
+                    JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
             }
 
             Debug.Assert(nextByte >= '0' && nextByte <= '9');
@@ -1495,12 +1402,19 @@ namespace System.Text.JsonLab
                 if (i < data.Length)
                 {
                     nextByte = data[i];
+                    if (delimiters.IndexOf(nextByte) != -1)
+                        goto Done;
+
                     if (nextByte != '.' && nextByte != 'E' && nextByte != 'e')
+#if UseInstrumented
+                        throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected '.' or 'E' or 'e'.", _lineNumber, _position);
+#else
                         JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
                     i++;
                 }
                 else
-                    return;
+                    goto Done;
             }
             else
             {
@@ -1510,10 +1424,14 @@ namespace System.Text.JsonLab
                     if ((uint)(nextByte - '0') > '9' - '0')
                         break;
                 }
-                if (i >= data.Length)
-                    return;
+                if (i >= data.Length || delimiters.IndexOf(nextByte) != -1)
+                    goto Done;
                 if (nextByte != '.' && nextByte != 'E' && nextByte != 'e')
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected '.' or 'E' or 'e'.", _lineNumber, _position);
+#else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
                 i++;
             }
 
@@ -1522,7 +1440,20 @@ namespace System.Text.JsonLab
             if (nextByte == '.')
             {
                 if (i >= data.Length)
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
+
+                nextByte = data[i];
+
+                if (delimiters.IndexOf(nextByte) != -1)
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
+                    JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
 
                 for (; i < data.Length; i++)
                 {
@@ -1530,24 +1461,45 @@ namespace System.Text.JsonLab
                     if ((uint)(nextByte - '0') > '9' - '0')
                         break;
                 }
-                if (i >= data.Length)
-                    return;
+                if (i >= data.Length || delimiters.IndexOf(nextByte) != -1)
+                    goto Done;
                 if (nextByte != 'E' && nextByte != 'e')
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected 'E' or 'e'.", _lineNumber, _position);
+#else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
                 i++;
             }
 
             Debug.Assert(nextByte == 'E' || nextByte == 'e');
 
             if (i >= data.Length)
+#if UseInstrumented
+                throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
                 JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
+
             nextByte = data[i];
             if (nextByte == '+' || nextByte == '-')
             {
                 i++;
                 if (i >= data.Length)
+#if UseInstrumented
+                    throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
+                nextByte = data[i];
             }
+
+            if (delimiters.IndexOf(nextByte) != -1)
+#if UseInstrumented
+                throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
+#else
+                JsonThrowHelper.ThrowJsonReaderException(ref this);
+#endif
 
             for (; i < data.Length; i++)
             {
@@ -1556,9 +1508,15 @@ namespace System.Text.JsonLab
                     break;
             }
 
-            if (i < data.Length)
+            if (i < data.Length && delimiters.IndexOf(nextByte) == -1)
+#if UseInstrumented
+                throw new JsonReaderException($"Invalid end of number. Last character read: '{(char)nextByte}'. Expected a delimiter.", _lineNumber, _position);
+#else
                 JsonThrowHelper.ThrowJsonReaderException(ref this);
-        }
 #endif
+
+            Done:
+            return data.Slice(0, i);
+        }
     }
 }