--- a/src/System.Text.JsonLab/System/Text/Json/Utf8JsonReaderInstrumented.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/Utf8JsonReaderInstrumented.cs
@@ -1390,7 +1390,7 @@ namespace System.Text.JsonLab
 #endif
 
                 nextByte = data[i++];
-                if (delimiters.IndexOf(nextByte) != -1)
+                if ((uint)(nextByte - '0') > '9' - '0')
 #if UseInstrumented
                     throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
 #else
@@ -1449,9 +1449,9 @@ namespace System.Text.JsonLab
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
 
-                nextByte = data[i];
+                nextByte = data[i++];
 
-                if (delimiters.IndexOf(nextByte) != -1)
+                if ((uint)(nextByte - '0') > '9' - '0')
 #if UseInstrumented
                     throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
 #else
@@ -1464,7 +1464,7 @@ namespace System.Text.JsonLab
                     if ((uint)(nextByte - '0') > '9' - '0')
                         break;
                 }
-                if (i >= data.Length || delimiters.IndexOf(nextByte) != -1)
+                if (i >= data.Length || delimiters.IndexOf(data[i]) != -1)
                     goto Done;
                 if (nextByte != 'E' && nextByte != 'e')
 #if UseInstrumented
@@ -1494,10 +1494,10 @@ namespace System.Text.JsonLab
 #else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-                nextByte = data[i];
+                nextByte = data[i++];
             }
 
-            if (delimiters.IndexOf(nextByte) != -1)
+            if ((uint)(nextByte - '0') > '9' - '0')
 #if UseInstrumented
                 throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
 #else
@@ -1511,7 +1511,7 @@ namespace System.Text.JsonLab
                     break;
             }
 
-            if (i < data.Length && delimiters.IndexOf(nextByte) == -1)
+            if (i < data.Length && delimiters.IndexOf(data[i]) == -1)
 #if UseInstrumented
                 throw new JsonReaderException($"Invalid end of number. Last character read: '{(char)nextByte}'. Expected a delimiter.", _lineNumber, _position);
 #else