--- a/src/System.Text.JsonLab/System/Text/Json/Utf8JsonReaderInstrumented.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/Utf8JsonReaderInstrumented.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+// KEEP THIS FILE IN-SYNC WITH JsonReader.cs, except for the define directive
 //TODO: Add multi-segment support
 #define UseInstrumented
 
@@ -1378,10 +1379,11 @@ namespace System.Text.JsonLab
             ReadOnlySpan<byte> delimiters = JsonConstants.Delimiters;
 
             int i = 0;
-            byte nextByte = data[i++];
+            byte nextByte = data[i];
 
             if (nextByte == '-')
             {
+                i++;
                 if (i >= data.Length)
 #if UseInstrumented
                     throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
@@ -1389,7 +1391,7 @@ namespace System.Text.JsonLab
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
 
-                nextByte = data[i++];
+                nextByte = data[i];
                 if ((uint)(nextByte - '0') > '9' - '0')
 #if UseInstrumented
                     throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
@@ -1402,6 +1404,7 @@ namespace System.Text.JsonLab
 
             if (nextByte == '0')
             {
+                i++;
                 if (i < data.Length)
                 {
                     nextByte = data[i];
@@ -1414,13 +1417,13 @@ namespace System.Text.JsonLab
 #else
                         JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-                    i++;
                 }
                 else
                     goto Done;
             }
             else
             {
+                i++;
                 for (; i < data.Length; i++)
                 {
                     nextByte = data[i];
@@ -1435,36 +1438,34 @@ namespace System.Text.JsonLab
 #else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-                i++;
             }
 
             Debug.Assert(nextByte == '.' || nextByte == 'E' || nextByte == 'e');
 
             if (nextByte == '.')
             {
+                i++;
                 if (i >= data.Length)
 #if UseInstrumented
                     throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
 #else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-
-                nextByte = data[i++];
-
+                nextByte = data[i];
                 if ((uint)(nextByte - '0') > '9' - '0')
 #if UseInstrumented
                     throw new JsonReaderException($"Invalid number. Last character read: '{(char)nextByte}'. Expected a digit.", _lineNumber, _position);
 #else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-
+                i++;
                 for (; i < data.Length; i++)
                 {
                     nextByte = data[i];
                     if ((uint)(nextByte - '0') > '9' - '0')
                         break;
                 }
-                if (i >= data.Length || delimiters.IndexOf(data[i]) != -1)
+                if (i >= data.Length || delimiters.IndexOf(nextByte) != -1)
                     goto Done;
                 if (nextByte != 'E' && nextByte != 'e')
 #if UseInstrumented
@@ -1472,10 +1473,10 @@ namespace System.Text.JsonLab
 #else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-                i++;
             }
 
             Debug.Assert(nextByte == 'E' || nextByte == 'e');
+            i++;
 
             if (i >= data.Length)
 #if UseInstrumented
@@ -1494,7 +1495,7 @@ namespace System.Text.JsonLab
 #else
                     JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-                nextByte = data[i++];
+                nextByte = data[i];
             }
 
             if ((uint)(nextByte - '0') > '9' - '0')
@@ -1503,7 +1504,7 @@ namespace System.Text.JsonLab
 #else
                 JsonThrowHelper.ThrowJsonReaderException(ref this);
 #endif
-
+            i++;
             for (; i < data.Length; i++)
             {
                 nextByte = data[i];
@@ -1511,7 +1512,7 @@ namespace System.Text.JsonLab
                     break;
             }
 
-            if (i < data.Length && delimiters.IndexOf(data[i]) == -1)
+            if (i < data.Length && delimiters.IndexOf(nextByte) == -1)
 #if UseInstrumented
                 throw new JsonReaderException($"Invalid end of number. Last character read: '{(char)nextByte}'. Expected a delimiter.", _lineNumber, _position);
 #else