Instead of using IndexOfAny(JsonConstants.Delimiters) to find the end of a number and then validating the number, read through the number and validate as we go.

Resulted in significant improvement (up to 40% in the best case) depending on how many numbers existed within the payload.

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17754
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009460
  [Host]     : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT
  Job-SSUQKN : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT

IterationCount=10  WarmupCount=5  

```

**BEFORE:**

|                               Method | IsDataCompact |       TestCase |          Mean |         Error |         StdDev | Allocated |
|------------------------------------- |-------------- |--------------- |--------------:|--------------:|---------------:|----------:|
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |      **BasicJson** |     **658.91 ns** |     **46.162 ns** |     **30.5334 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **BasicLargeNum** |     **599.76 ns** |     **11.982 ns** |      **7.1301 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |     **HelloWorld** |      **70.66 ns** |      **1.324 ns** |      **0.6926 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |      **Json400KB** | **889,774.57 ns** | **22,262.378 ns** | **14,725.1855 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **LotsOfNumbers** |   **3,888.03 ns** |     **74.180 ns** |     **44.1435 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** | **SpecialNumForm** |   **1,611.65 ns** |    **110.629 ns** |     **73.1742 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |      **BasicJson** |     **451.48 ns** |     **21.443 ns** |     **14.1832 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicLargeNum** |     **734.43 ns** |     **24.650 ns** |     **16.3044 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |     **HelloWorld** |      **61.58 ns** |      **2.746 ns** |      **1.8165 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |      **Json400KB** | **665,780.11 ns** | **27,569.075 ns** | **16,405.9199 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **LotsOfNumbers** |   **3,291.76 ns** |    **150.555 ns** |     **99.5826 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **SpecialNumForm** |   **1,450.84 ns** |     **39.853 ns** |     **20.8439 ns** |       **0 B** |

**AFTER:**

|                               Method | IsDataCompact |       TestCase |          Mean |          Error |        StdDev | Allocated |
|------------------------------------- |-------------- |--------------- |--------------:|---------------:|--------------:|----------:|
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |      **BasicJson** |     **562.92 ns** |     **10.9657 ns** |     **7.2532 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **BasicLargeNum** |     **463.39 ns** |      **6.1524 ns** |     **3.2178 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |     **HelloWorld** |      **72.14 ns** |      **0.4956 ns** |     **0.2949 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |      **Json400KB** | **815,310.31 ns** | **13,052.7690 ns** | **8,633.5988 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **LotsOfNumbers** |   **2,821.09 ns** |     **27.3199 ns** |    **16.2576 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** | **SpecialNumForm** |   **1,295.33 ns** |     **25.3528 ns** |    **15.0870 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |      **BasicJson** |     **336.49 ns** |      **7.1159 ns** |     **3.7218 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicLargeNum** |     **467.04 ns** |     **12.1527 ns** |     **6.3561 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |     **HelloWorld** |      **63.68 ns** |      **7.8392 ns** |     **4.6650 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |      **Json400KB** | **590,439.59 ns** |  **7,408.7881 ns** | **4,900.4548 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **LotsOfNumbers** |   **1,934.43 ns** |     **23.8542 ns** |    **12.4763 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **SpecialNumForm** |   **1,205.45 ns** |     **88.8544 ns** |    **58.7717 ns** |       **0 B** |